Moved multi-exp acc loop into iterator fold method
